@atomic-reactor/reactium-sdk-core@0.0.60
Now attempts to correct "Cannot update a component from inside the function body of a different component" error occuring in React >=16.13.1 on useRegisterHandle() without breaking components that rely on early component registration.

There is still at least one case where useHandle() subscribers will not have the handle (if the user component renders before the component with the handle), but in such a case, at least there will be a sane empty object default, if no default is provided.